% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_score.R
\name{calc_score}
\alias{calc_score}
\title{Calculate cell type score of genes}
\usage{
calc_score(specificity, cell_type)
}
\arguments{
\item{specificity}{a matrix gotten from calc_specificity function (without NA) with row names,
rows denoting genes.}

\item{cell_type}{a vector of cell type annotation, whose dimension equals to ncol(datExpr).}
}
\value{
A dataframe including cell type score for each gene, row denoting genes.
}
\description{
Calculation of score for each gene using cell type specificity.
}
\examples{
{}
}
