% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr_disease_module.R
\name{constr_disease_module}
\alias{constr_disease_module}
\title{Construct cell type-specific disease gene module}
\usage{
constr_disease_module(
  Cts_list,
  disease_gene,
  itera = 1000,
  returnAllmodules = FALSE,
  padjCutoff = 0.1
)
}
\arguments{
\item{Cts_list}{a list, gotten from constr_specific_net function, containing cell type-specific gene network, a two-column matrix, character or numeric, with each row defining one edge,
and genescore, a dataframe of cell type scores of genes in the cell type-specific gene network.}

\item{disease_gene}{a vector of disease-associated genes. The keytype of disease_gene must be consistent with the one of network.}

\item{itera}{a number (a positive integer) of resampling iterations (by default 1000).}

\item{returnAllmodules}{a boolean value determining if all candidate cell type-specific disease gene modules should be return or not (by default FALSE, only return the significant ones).}

\item{padjCutoff}{a cutoff of fdr-adjusted P value.}
}
\value{
A list containing disease_module, cell type-specific disease gene module, each row defining one edge;
gene_score, cell type scores of genes in the cell type-specific disease gene module; Pvalue, significance of the disease module
and fdr-adjusted P value.
}
\description{
Identification of cell type-specific disease gene module from cell type-specific gene network,
              using cell type scores of genes and disease-associated genes.
}
\examples{
{}
}
