% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr_specific_net.R
\name{constr_specific_net}
\alias{constr_specific_net}
\title{Construct cell type-specific gene network}
\usage{
constr_specific_net(net, genescore, sthre = 0)
}
\arguments{
\item{net}{a referenced gene-gene interaction network, a two-column matrix, character or numeric, each row defining one edge.}

\item{genescore}{a dataframe of cell type scores of genes, gotten from calc_score function.}

\item{sthre}{a cutoff of cell type scores used for getting cell type-specific genes (by default 0).}
}
\value{
A list containing specific_net, cell type-specific gene network each row defining one edge;
and gene_score, cell type score for each gene in the cell type-specific gene network.
}
\description{
Construction of cell type-specific gene network from a referenced gene-gene interaction network,
              using cell type scores of genes.
}
\examples{
{}
}
